show fractional amounts of users balance
