+++ src/ui/views/TokenDetail/TokenValue.tsx
@@ -1,6 +1,7 @@
 import React from 'react';
 
 import { formatPrice } from '@/shared/utils/formatTokenValue';
+import { numberWithCommas } from '@/ui/utils';
 
 interface TokenPriceProps {
   value: string;
@@ -16,25 +17,11 @@ export const TokenValue: React.FC<TokenPriceProps> = ({
   prefix = '',
   postFix = '',
 }) => {
+  console.log('value', value, prefix, postFix);
+
   if (!value || value === '0' || value === '') {
     return <span className={className}>{''}</span>;
   }
-
-  const numberWithCommas = (x: string) => {
-    // Split string into integer and decimal parts (if any)
-    const parts = x.split('.');
-    const integerPart = parts[0];
-
-    // Only add commas if the integer part is between 4 and 6 digits
-    if (integerPart.length >= 4 && integerPart.length <= 6) {
-      // Add commas every 3 digits from the right
-      const withCommas = integerPart.replace(/\B(?=(\d{3})+(?!\d))/g, ',');
-      // Reconstruct number with decimal part if it exists
-      return parts.length > 1 ? `${withCommas}.${parts[1]}` : withCommas;
-    }
-
-    return x;
-  };
 
   const { formattedPrice } = formatPrice(value);
   const { leadingPart, zeroPart, endingPart } = formattedPrice;
